Makes pagination a reusable component
