+++ src/components/coins/FullCoinList.tsx
@@ -1,22 +1,15 @@
-import React, { useState } from 'react';
-import CoinListItem from './CoinListItem';
+import React from 'react';
 import LoadingPanel from '../ui/LoadingPanel';
-import PageNumbers from '../ui/Pagination';
-import { Coin } from '../api/coins';
-import { useGlobalStateContext } from '../../utils/GlobalStateProvider';
 import useCoinDataQuery from '../hooks/useCoinDataQuery';
-
-
-
+import PaginatedList from '../ui/pagination/PaginatedList';
+import CoinListItem from './CoinListItem';
+import { useGlobalStateContext } from '../../utils/GlobalStateProvider';
 
 const FullCoinList = () => {
-    const [pageIndex, setPageIndex] = useState<number>(0);
-    const { state } = useGlobalStateContext();
     const { isLoading, data } = useCoinDataQuery();
+    const { state } = useGlobalStateContext();
 
     if (isLoading) return <LoadingPanel />;
-
-    let currentIndex = 0;
 
     const coinData = Object.entries(data?.Data || {}).map(([key, coin]) => ({
         id: key,
@@ -25,29 +18,18 @@ const FullCoinList = () => {
         code: key,
     }));
 
-    const pages = coinData.reduce((acc: Coin[][], cur: Coin, index) => {
-        if (!acc[currentIndex]) acc.push([]);
-        acc[currentIndex].push(cur);
-    
-        if ((index + 1) % 50 === 0) currentIndex += 1;
-        return acc;
-    }, []);
-
     return (
         <div className="coin-list">
-            {pages[pageIndex].map(coin => (
-                <CoinListItem
-                    coin={coin}
-                    isCurrencyDollar={state.isCurrencyDollar}
-                    key={coin.id}
-                    isFavourite={state.activeCoinCodes.includes(coin.code)}
-                />
-            ))}
-
-            <PageNumbers
-                currentPageIndex={pageIndex}
-                totalPages={pages.length}
-                onChangePage={(newPage) => setPageIndex(newPage)}
+            <PaginatedList
+                data={coinData}
+                renderItem={({ item }) => (
+                    <CoinListItem
+                        coin={item}
+                        isCurrencyDollar={state.isCurrencyDollar}
+                        key={item.id}
+                        isFavourite={state.activeCoinCodes.includes(item.code)}
+                    />
+                )}
             />
         </div>
     );
